Update mv_multiply
Depending on the size of the multivector(-batches) a, b this commit adds a case distinction in order to be able to process larger batch-sizes of multivectors a (exclusive) or b. This makes sense when considering the theory (formula, non-commutativity) and use cases of the geometric product. In praxis the multivector-batch can be a resp. b.